Added: New, more Unity like, icon.
This icon suites the Unity optics much better, although it is still not
perfect. Thx to @vovkkk for the SVG version, for which I postprocessed
the PNG export a bit in GIMP to make it look less blurry. Still: If
someone feels to make a better icon, feel welcome to suggest one.