Improving detection of external scripts.
If the theme-changing scripts are not installed in the expected
location, then don't show the menu items.

Searches for `edit-user-config` either in the system-wide location or in
the same directory. If none is found, don't show the menu item.

Also moving some class variables to intance variables.

Also fixing whitespace according to PEP-8.